% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Academic theme in addition to default ggplot2 themes}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{A positive integer}
}
\value{
A plot without the theme added when passed with ggplot2 (centralized plot title, gridlines removed)
}
\description{
Academic theme in addition to default ggplot2 themes
}
\examples{
library(ggplot2)
p<-ggplot(mtcars)+geom_point(aes(x=mpg,y=hp))+labs(title="plot_title")
p+theme_academic
}
